large last minute patch, reviewed by Dirk Mueller:
* removed obsolete ifdef'ed Qt 2 code
* fixed header display, fonts
* fixed charset/encoding handling
* fixed default colors / enabled alternate colors
  in the header view, like in kmail / default fonts
  improved
* the remove trailing newlines options works again
* fixed attachment display
* anchors are clickable even if the user has scrolled
  down a bit (#38212)
* fixed anchor handling in general (#36365)
* article beautification works again (#38982)
* cleaned up composer toolbar
* fixed rot-13 and other special functions of the composer
  (#39073)
* applied patch for null-character bug (#33058)
* added two missing i18n() calls reported by Thomas Diehl
* fixed DnD behaviour (#38176)
* fixed unsubscribe bug (#38660)

knode should be useable now....
new version number: 0.7.1

svn path=/trunk/kdenetwork/knode/; revision=142448

+++ knode/kncleanup.cpp
@@ -220,7 +220,7 @@ void KNCleanUp::compactFolder(KNFolder *f)
       if(f->m_boxFile.at(art->startOffset())) {
         ts << "From aaa@aaa Mon Jan 01 00:00:00 1997\n";
         art->setStartOffset(newMBoxFile.at());
-        while(f->m_boxFile.at() < art->endOffset())
+        while(f->m_boxFile.at() < (uint)art->endOffset())
           ts << f->m_boxFile.readLineWnewLine();
         art->setEndOffset(newMBoxFile.at());
         newMBoxFile.putch('\n');
